--- src/main/java/com/example/HomeSaveHome/repository/UserRepository.java
@@ -1,4 +1,0 @@
-package com.example.HomeSaveHome.repository;
-
-public interface UserRepository {
-}
