+++ src/main/java/com/example/HomeSaveHome/user/repository/UserRepository.java
@@ -1,0 +1,4 @@
+package com.example.HomeSaveHome.user.repository;
+
+public interface UserRepository {
+}
