+++ src/main/java/com/example/HomeSaveHome/user/security/SecurityConfig.java
@@ -1,4 +1,4 @@
-package com.example.HomeSaveHome.security;
+package com.example.HomeSaveHome.user.security;
 
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
